refactor: Extract style config
