+++ config/webpack.config.base.js
@@ -2,16 +2,12 @@
 
 const path = require('path')
 const ScriptExtHtmlWebpackPlugin = require('script-ext-html-webpack-plugin')
-const PostCSSAssetsPlugin = require('postcss-assets-webpack-plugin')
-const sortCSSmq = require('sort-css-media-queries')
 
-const { extractor, production } = require('./webpack.vars')
-const SRC_DIR = path.resolve(__dirname, '../src')
+const { production, SRC_DIR } = require('./webpack.vars')
 const pkg = require(path.resolve(__dirname, '../package.json'))
 const webpack = require('webpack')
 const DuplicatePackageCheckerPlugin = require('duplicate-package-checker-webpack-plugin');
 const StatsPlugin = require('stats-webpack-plugin')
-const MiniCssExtractPlugin = require('mini-css-extract-plugin')
 
 module.exports = {
   output: {
@@ -52,48 +48,6 @@ module.exports = {
           cacheDirectory: true
         }
       },
-      {
-        include: SRC_DIR,
-        test: /\.styl$/,
-        use: [
-          {
-            loader: MiniCssExtractPlugin.loader
-          }, {
-            loader: 'css-loader',
-            options: {
-              importLoaders: 2,
-              modules: true,
-              sourceMap: true,
-              localIdentName: '[name]_[local]_[hash:base64:5]'
-            }
-          }, {
-            loader: 'stylus-loader',
-            options: {
-              sourceMap: true
-            }
-          }]
-      },
-      {
-        test: /\.css$/,
-        use: [
-          {
-            loader: MiniCssExtractPlugin.loader
-          },
-          {
-            loader: 'css-loader',
-            options: {
-              importLoaders: 1
-            }
-          },
-          {
-            loader: 'postcss-loader',
-            options: {
-              sourceMap: true,
-              plugins: loader => [require('autoprefixer')()]
-            }
-          }
-        ]
-      },
       // Fonts
       {
         test: /\.woff2?$/,
@@ -113,28 +67,7 @@ module.exports = {
     new ScriptExtHtmlWebpackPlugin({
       defaultAttribute: 'defer'
     }),
-    new MiniCssExtractPlugin({
-      // Options similar to the same options in webpackOptions.output
-      // both options are optional
-      filename: `app${production ? '.[hash].min' : ''}.css`
-    }),
     new DuplicatePackageCheckerPlugin({ verbose: true }),
-    new PostCSSAssetsPlugin({
-      test: /\.css$/,
-      plugins: [
-        require('autoprefixer')(['last 2 versions']),
-        require('css-mqpacker')({sort: sortCSSmq}),
-        require('postcss-discard-duplicates'),
-        require('postcss-discard-empty')
-      ].concat(
-        production
-          ? require('csswring')({
-              preservehacks: true,
-              removeallcomments: true
-            })
-          : []
-      )
-    })
   ].concat(process.env.WEBPACK_STATS_FILE
     ? [ new StatsPlugin(process.env.WEBPACK_STATS_FILE, {
         chunkModules: true,
