+++ config/webpack.vars.js
@@ -1,6 +1,6 @@
 'use strict'
 
-const MiniCssExtractPlugin = require('mini-css-extract-plugin')
+const path = require('path')
 
 const getTarget = () => {
   try {
@@ -14,11 +14,13 @@ const production = /:production$/.test(process.env.NODE_ENV)
 const target = getTarget()
 const hotReload = !!process.env.HOT_RELOAD
 const skin = process.env.SKIN
+const SRC_DIR = path.resolve(__dirname, '../src')
 
 module.exports = {
   production: production,
   target: target,
   hotReload,
   analyze: process.env.WEBPACK_ANALYZE,
-  skin
+  skin,
+  SRC_DIR
 }
