+++ webpack.config.js
@@ -11,6 +11,7 @@ const {
 const common = merge(
   require('./config/webpack.config.base'),
   require('./config/webpack.config.disable-contexts'),
+  require('./config/webpack.config.styles'),
   require('./config/webpack.config.cozy-ui'),
   require('./config/webpack.config.pictures'),
   require('./config/webpack.config.vendors'),
